pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "weave-cli"
version = "0.1.0"
description = "Declarative agent orchestration framework inspired by Terraform"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Weave Team", email = "hello@weave.dev"}
]
keywords = ["ai", "agents", "orchestration", "cli", "workflow"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.7.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "networkx>=3.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.12.0",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
    "types-pyyaml>=6.0.12",
]

[project.scripts]
weave = "weave.cli.app:main"

[project.urls]
Homepage = "https://github.com/weave/weave-cli"
Documentation = "https://docs.weave.dev"
Repository = "https://github.com/weave/weave-cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
